requirements.txt
# Core
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Vision (SEE stage)
opencv-python>=4.9.0
mediapipe>=0.10.0

# Embeddings (local, real)
sentence-transformers>=3.0.0

# Vector store
sqlite-vec>=0.1.1

# Observability / Dashboard
psutil>=5.9.0
streamlit>=1.35.0
plotly>=5.22.0
pandas>=2.0.0

# MLX fine-tuning (Apple Silicon)
mlx>=0.16.0
mlx-lm>=0.18.0
mlx-vlm>=0.3.0

# Utilities
numpy>=1.26.0
